refactor: move org logo logic to orgfunctions and endpoints to user_org_api
