+++ ddpui/routes.py
@@ -26,8 +26,6 @@ from ddpui.api.public_api import public_router
 from ddpui.api.report_api import report_router
 from ddpui.api.metric_api import metric_router
 from ddpui.api.kpi_api import kpi_router
-from ddpui.api.org_logo_api import org_logo_router
-
 
 src_api = NinjaAPI(
     urls_namespace="api",
@@ -92,7 +90,6 @@ public_router.tags = ["Public"]
 report_router.tags = ["Reports"]
 metric_router.tags = ["Metrics"]
 kpi_router.tags = ["KPIs"]
-org_logo_router.tags = ["OrgLogo"]
 
 # mount all the module routes
 src_api.add_router("/api/airbyte/", airbyte_router)
@@ -116,7 +113,6 @@ src_api.add_router("/api/filters/", filter_router)
 src_api.add_router("/api/reports/", report_router)
 src_api.add_router("/api/metrics/", metric_router)
 src_api.add_router("/api/kpis/", kpi_router)
-src_api.add_router("/api/org/logo/", org_logo_router)
 
 # Public API without authentication
 public_api = NinjaAPI(
